+++ divio_cli/cli.py
@@ -265,34 +265,34 @@ def application_list(obj, grouped, as_json):
 
 
 @app.command(name="deploy")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @allow_remote_id_override
 @click.pass_obj
-def application_deploy(obj, remote_id, stage):
+def application_deploy(obj, remote_id, environment):
     """Deploy application."""
-    obj.client.deploy_application_or_get_progress(remote_id, stage)
+    obj.client.deploy_application_or_get_progress(remote_id, environment)
 
 
 @app.command(name="deploy-log")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @allow_remote_id_override
 @click.pass_obj
-def application_deploy_log(obj, remote_id, stage):
+def application_deploy_log(obj, remote_id, environment):
     """View last deployment log."""
-    deploy_log = obj.client.get_deploy_log(remote_id, stage)
+    deploy_log = obj.client.get_deploy_log(remote_id, environment)
     if deploy_log:
         echo_large_content(deploy_log, ctx=obj)
     else:
         click.secho(
             "Environment with the name '{}' does not exist, no log available.".format(
-                stage
+                environment
             ),
             fg="yellow",
         )
 
 
 @app.command(name="logs")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @click.option(
     "--tail", "tail", default=False, is_flag=True, help="Tail the output."
 )
@@ -301,18 +301,18 @@ def application_deploy_log(obj, remote_id, stage):
 )
 @allow_remote_id_override
 @click.pass_obj
-def application_logs(obj, remote_id, stage, tail, utc):
+def application_logs(obj, remote_id, environment, tail, utc):
     """View logs."""
-    obj.client.show_log(remote_id, stage, tail, utc)
+    obj.client.show_log(remote_id, environment, tail, utc)
 
 
 @app.command(name="ssh")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @allow_remote_id_override
 @click.pass_obj
-def application__ssh(obj, remote_id, stage):
+def application__ssh(obj, remote_id, environment):
     """Establish SSH connection."""
-    obj.client.ssh(remote_id, stage)
+    obj.client.ssh(remote_id, environment)
 
 
 @app.command(name="configure")
@@ -343,14 +343,14 @@ def application_down():
 
 
 @app.command(name="open")
-@click.argument("stage", default="")
+@click.argument("environment", default="")
 @allow_remote_id_override
 @click.pass_obj
-def application_open(obj, remote_id, stage):
+def application_open(obj, remote_id, environment):
     """Open local or cloud applications in a browser."""
-    if stage:
+    if environment:
         open_application_cloud_site(
-            obj.client, application_id=remote_id, stage=stage
+            obj.client, application_id=remote_id, environment=environment
         )
     else:
         localdev.open_application()
@@ -383,7 +383,7 @@ def application_update(obj, strict):
 @app.command(name="env-vars")
 @click.option(
     "-s",
-    "--stage",
+    "--environment",
     default="test",
     type=str,
     help="Manage the cloud application's environment variables.",
@@ -427,7 +427,7 @@ def application_update(obj, strict):
 def environment_variables(
     obj,
     remote_id,
-    stage,
+    environment,
     show_all_vars,
     as_json,
     get_vars,
@@ -439,18 +439,20 @@ def environment_variables(
 
     WARNING: This command is experimental and may change in a future release.
     """
-    stage = stage.lower()
+    environment = environment.lower()
     if set_vars or unset_vars:
         set_vars = dict(set_vars)
         data = obj.client.set_custom_environment_variables(
             website_id=remote_id,
-            stage=stage,
+            environment=environment,
             set_vars=set_vars,
             unset_vars=unset_vars,
         )
     else:
         data = obj.client.get_environment_variables(
-            website_id=remote_id, stage=stage, custom_only=not show_all_vars
+            website_id=remote_id,
+            environment=environment,
+            custom_only=not show_all_vars,
         )
         if get_vars:
             data = {
@@ -475,7 +477,7 @@ def app_status():
 @click.argument("slug")
 @click.option(
     "-s",
-    "--stage",
+    "--environment",
     default="test",
     help="Specify environment from which media and content data will be pulled.",
 )
@@ -499,7 +501,7 @@ def app_status():
     help="Skip system test before setting up the application.",
 )
 @click.pass_obj
-def application_setup(obj, slug, stage, path, overwrite, skip_doctor):
+def application_setup(obj, slug, environment, path, overwrite, skip_doctor):
     """Set up a development environment for a Divio application."""
     if not skip_doctor and not check_requirements_human(
         config=obj.client.config, silent=True
@@ -513,7 +515,7 @@ def application_setup(obj, slug, stage, path, overwrite, skip_doctor):
         sys.exit(1)
 
     localdev.create_workspace(
-        obj.client, slug, stage, path, overwrite, obj.zone
+        obj.client, slug, environment, path, overwrite, obj.zone
     )
 
 
@@ -529,11 +531,11 @@ def application_pull():
     default=False,
     help="Keep the temporary file with the data.",
 )
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @click.argument("prefix", default=localdev.DEFAULT_SERVICE_PREFIX)
 @allow_remote_id_override
 @click.pass_obj
-def pull_db(obj, remote_id, stage, prefix, keep_tempfile):
+def pull_db(obj, remote_id, environment, prefix, keep_tempfile):
     """
     Pull database the Divio cloud environment.
     """
@@ -545,7 +547,7 @@ def pull_db(obj, remote_id, stage, prefix, keep_tempfile):
 
     localdev.ImportRemoteDatabase(
         client=obj.client,
-        stage=stage,
+        environment=environment,
         prefix=prefix,
         remote_id=remote_id,
         db_type=db_type,
@@ -555,14 +557,16 @@ def pull_db(obj, remote_id, stage, prefix, keep_tempfile):
 
 
 @application_pull.command(name="media")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @allow_remote_id_override
 @click.pass_obj
-def pull_media(obj, remote_id, stage):
+def pull_media(obj, remote_id, environment):
     """
     Pull media files from the Divio cloud environment.
     """
-    localdev.pull_media(obj.client, stage=stage, remote_id=remote_id)
+    localdev.pull_media(
+        obj.client, environment=environment, remote_id=remote_id
+    )
 
 
 @app.group(name="push")
@@ -571,7 +575,7 @@ def application_push():
 
 
 @application_push.command(name="db")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @click.option(
     "-d",
     "--dumpfile",
@@ -588,7 +592,7 @@ def application_push():
 @click.argument("prefix", default=localdev.DEFAULT_SERVICE_PREFIX)
 @allow_remote_id_override
 @click.pass_obj
-def push_db(obj, remote_id, prefix, stage, dumpfile, noinput):
+def push_db(obj, remote_id, prefix, environment, dumpfile, noinput):
     """
     Push database to the Divio cloud environment..
     """
@@ -598,24 +602,30 @@ def push_db(obj, remote_id, prefix, stage, dumpfile, noinput):
     db_type = utils.get_db_type(prefix, path=application_home)
     if not dumpfile:
         if not noinput:
-            click.secho(messages.PUSH_DB_WARNING.format(stage=stage), fg="red")
+            click.secho(
+                messages.PUSH_DB_WARNING.format(environment=environment),
+                fg="red",
+            )
             if not click.confirm("\nAre you sure you want to continue?"):
                 return
         localdev.push_db(
             client=obj.client,
-            stage=stage,
+            environment=environment,
             remote_id=remote_id,
             prefix=prefix,
             db_type=db_type,
         )
     else:
         if not noinput:
-            click.secho(messages.PUSH_DB_WARNING.format(stage=stage), fg="red")
+            click.secho(
+                messages.PUSH_DB_WARNING.format(environment=environment),
+                fg="red",
+            )
             if not click.confirm("\nAre you sure you want to continue?"):
                 return
         localdev.push_local_db(
             obj.client,
-            stage=stage,
+            environment=environment,
             dump_filename=dumpfile,
             website_id=remote_id,
             prefix=prefix,
@@ -623,7 +633,7 @@ def push_db(obj, remote_id, prefix, stage, dumpfile, noinput):
 
 
 @application_push.command(name="media")
-@click.argument("stage", default="test")
+@click.argument("environment", default="test")
 @click.option(
     "--noinput",
     is_flag=True,
@@ -633,17 +643,20 @@ def push_db(obj, remote_id, prefix, stage, dumpfile, noinput):
 @click.argument("prefix", default=localdev.DEFAULT_SERVICE_PREFIX)
 @allow_remote_id_override
 @click.pass_obj
-def push_media(obj, remote_id, prefix, stage, noinput):
+def push_media(obj, remote_id, prefix, environment, noinput):
     """
     Push database to the Divio cloud environment..
     """
 
     if not noinput:
-        click.secho(messages.PUSH_MEDIA_WARNING.format(stage=stage), fg="red")
+        click.secho(
+            messages.PUSH_MEDIA_WARNING.format(environment=environment),
+            fg="red",
+        )
         if not click.confirm("\nAre you sure you want to continue?"):
             return
     localdev.push_media(
-        obj.client, stage=stage, remote_id=remote_id, prefix=prefix
+        obj.client, environment=environment, remote_id=remote_id, prefix=prefix
     )