+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Unreleased
 * Support ``-h`` as a shorthand for the ``--help`` option.
 * Support ``-P`` to deactivate pagers when viewing potentially larger amount of data.
 * Changed output of some error or warning messages from stdout to stderr.
+* Renamed the STAGE param into ENVIRONMENT in CLI commands
 
 3.15.7 (2022-09-23)
 -------------------